Change xml attributes order in test data.
Reason: tests in python <= 3.7.

ElementTree up to 3.7 always sorts attributes lexicographically.
Since 3.8 user specified order is used.

Semantical xml comparison can not be used as Concept Explorer works
with cex xml in nonstadrad way - order of elements IS important
while order of attributes IS NOT important.

Another way (possible cleaner?) to fix thist would be to write
xml comparator which takes account only of elements order.
Would it be overkill solution?

Or drop python <= 3.7 support. Again overkill?

Or make differnet set of tests for <=3.7 AND FOR > 3.7.
 Only difference would be the ordering of attributes.